PHNX-142 | Передвижение игрока в катсцене
> MovingController сделан утилитарным и теперь может использоваться для передвижения игрока
> Фикс скипа катсцен. Больше не очищается ссылка на skipTrigger

+++ assets/scripts/characters/player/PegasusSmash.cs
@@ -16,7 +16,7 @@ public class PegasusSmash : Area
         if (body is Player) return;
         if (!player.MaySmash || player.GetSpeed() <= 6) return;
 
-        var tempDamage = player.GetSpeed() * 3;
+        var tempDamage = (int)player.GetSpeed() * 3;
 
         switch (body)
         {
